// src/out/models/CreateUserInput.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ユーザー
 * ユーザーCRUD操作をサーバーレス(aws)で構築する
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: apiteam@swagger.io
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 *
 * @export
 * @interface CreateUserInput
 */
export interface CreateUserInput {
    /**
     *
     * @type {string}
     * @memberof CreateUserInput
     */
    firstName?: string;
    /**
     *
     * @type {string}
     * @memberof CreateUserInput
     */
    lastName?: string;
}

export function CreateUserInputFromJSON(json: any): CreateUserInput {
    return CreateUserInputFromJSONTyped(json, false);
}

export function CreateUserInputFromJSONTyped(json: any, ignoreDiscriminator: boolean): CreateUserInput {
    if (json === undefined || json === null) {
        return json;
    }
    return {
        firstName: !exists(json, 'firstName') ? undefined : json['firstName'],
        lastName: !exists(json, 'lastName') ? undefined : json['lastName'],
    };
}

export function CreateUserInputToJSON(value?: CreateUserInput | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        firstName: value.firstName,
        lastName: value.lastName,
    };
}